LayoutTests/sputnik/Conformance/12_Statement/12.8_The_break_Statement/S12.8_A1_T2-expected.txt

CONSOLE MESSAGE: line 81: SyntaxError: Label 'LABEL' is not defined
S12.8_A1_T2

PASS Expected parsing failure

TEST COMPLETE
